+++ OpenAI.SDK/EndpointProviders/OpenAiEndpointProvider.cs
@@ -1,28 +1,5 @@
-﻿namespace OpenAI.GPT3
+﻿namespace OpenAI.GPT3.EndpointProviders
 {
-    internal interface IOpenAiEndpointProvider
-    {
-        string ModelRetrieve(string model);
-        string CompletionCreate();
-        string EditCreate();
-        string ModelsList();
-        string FilesList();
-        string FilesUpload();
-        string FileDelete(string fileId);
-        string FileRetrieve(string fileId);
-        string FineTuneCreate();
-        string FineTuneList();
-        string FineTuneRetrieve(string fineTuneId);
-        string FineTuneCancel(string fineTuneId);
-        string FineTuneListEvents(string fineTuneId);
-        string FineTuneDelete(string fineTuneId);
-        string EmbeddingCreate();
-        string ModerationCreate();
-        string ImageCreate();
-        string ImageEditCreate();
-        string ImageVariationCreate();
-    }
-
     internal class OpenAiEndpointProvider : IOpenAiEndpointProvider
     {
         private readonly string _apiVersion;
